+++ pyodide-build/pyodide_build/common.py
@@ -1,9 +1,10 @@
+import contextlib
 import functools
 import os
 import subprocess
 import sys
 from pathlib import Path
-from typing import Iterable, Iterator
+from typing import Generator, Iterable, Iterator, Mapping
 
 import tomli
 from packaging.tags import Tag, compatible_tags, cpython_tags
@@ -257,3 +258,15 @@ def get_unisolated_packages():
     unisolated_packages.append("setuptools_rust")
     os.environ["UNISOLATED_PACKAGES"] = json.dumps(unisolated_packages)
     return unisolated_packages
+
+
+@contextlib.contextmanager
+def replace_env(build_env: Mapping[str, str]) -> Generator[None, None, None]:
+    old_environ = dict(os.environ)
+    os.environ.clear()
+    os.environ.update(build_env)
+    try:
+        yield
+    finally:
+        os.environ.clear()
+        os.environ.update(old_environ)
